split viewmodel into partial classes, group and format dialogsview
